test(inspector): scope localStorage mock to tests needing it, fix project-store isolation
