+++ apps/web/src/test/setup.ts
@@ -1,28 +1,5 @@
 import "@testing-library/jest-dom/vitest";
 
-if (typeof window.localStorage?.setItem !== "function") {
-  const store = new Map<string, string>();
-  const localStorageMock: Storage = {
-    get length() {
-      return store.size;
-    },
-    clear: () => store.clear(),
-    getItem: (key) => (store.has(key) ? store.get(key)! : null),
-    key: (index) => Array.from(store.keys())[index] ?? null,
-    removeItem: (key) => {
-      store.delete(key);
-    },
-    setItem: (key, value) => {
-      store.set(key, String(value));
-    },
-  };
-
-  Object.defineProperty(window, "localStorage", {
-    writable: true,
-    value: localStorageMock,
-  });
-}
-
 // Mock matchMedia for tests
 Object.defineProperty(window, "matchMedia", {
   writable: true,
